+++ PointCloud/Face.h
@@ -9,7 +9,7 @@ namespace pointcloud {
 	class Face {
 	public:
 		int shape_id;
-		std::vector<glm::dvec3> points;
+		std::vector<glm::vec3> points;
 		std::vector<std::vector<glm::dvec3>> triangles;
 		glm::dvec3 normal;
 		double d;
